apfsck: correct type width for s_data_blocks
I don't know how I missed this before, but sparse is complaining about
a call to le64_to_cpu() on the 32-bit nx_xp_data_blocks field of the
on-disk superblock.  Fix this and change the width of the corresponding
in-memory field.

Fixes: 96c80bbd0ec4 ("apfsck: check the address of ephemeral objects")

+++ apfsck/super.h
@@ -56,7 +56,7 @@ struct super_block {
 	u64 s_next_oid;	/* Next virtual object id to be used */
 	u32 s_max_vols; /* Maximum number of volumes allowed */
 	u64 s_data_base; /* Base address of the checkpoint data area */
-	u64 s_data_blocks; /* Number of blocks in the checkpoint data area */
+	u32 s_data_blocks; /* Number of blocks in the checkpoint data area */
 	u32 s_data_index; /* Index of first valid block in checkpoint data */
 	u32 s_data_len; /* Number of valid blocks in checkpoint data area */
